Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    unsigned int a;
-    scanf("%u",&a);
+    long unsigned int a;
+    scanf("%lu",&a);
     if(1&a){
         printf("Set");
     }
